services/Eci-catalog-service/requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.15
django-filter>=24.2
pymysql>=1.1.0
python-dotenv>=1.0.1
drf-spectacular>=0.27
drf-spectacular-sidecar>=2024.3.1   # optional, for Swagger UI assets
